# openapi/ortho-v1.yaml
openapi: 3.1.0
info:
  title: ORTHO-32 API
  version: 1.0.0
  description: |
    Canonical versioned REST/SSE contract all ORTHO clients talk to.
    All responses are wrapped in Envelope { id, type, timestamp, source, correlation_id, data }.
    cycles fields are integer architectural cycles, NEVER wall-clock ms or float.
    VerificationStatus has exactly 7 disjoint values, never interchangeable.
    Elevated scopes: hardware.reset, marketplace.install require explicit grant.
servers:
  - url: http://localhost:7032
    description: Local development (HTTP transport)
  - url: https://ortho.local:7032
    description: Remote HTTP/2 transport
  - url: ipc://ortho32-api.sock
    description: Local IPC transport (desktop co-located)
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Scoped capability token from POST /api/v1/auth/session
  schemas:
    Envelope:
      type: object
      required: [id, type, timestamp, source, correlation_id, data]
      properties:
        id: { type: string, format: uuid, example: "550e8400-e29b-41d4-a716-446655440000" }
        type: { type: string, example: "tensor.job.created" }
        timestamp: { type: string, format: date-time }
        source: { type: string, example: "ortho32-api" }
        correlation_id: { type: string, example: "req-123" }
        data: { description: "typed payload" }
    Cycles:
      type: integer
      minimum: 0
      description: "Architectural cycle count - integer. NEVER float, NEVER wall-clock ms."
      example: 1024
...
    VerificationStatus:
      type: string
      enum: [VERIFIED, CROSS_VERIFIED, TESTED, OBSERVED, ASSUMED, UNVERIFIED, FAILED]
      description: "Exactly 7 disjoint values. NEVER interchangeable."
...
    TensorJob:
      type: object
      required: [id, status, cycles, createdAt]
      properties:
        id: { type: string }
        status: { type: string, enum: [queued, running, completed, failed] }
        model: { type: string }
        cycles: { $ref: "#/components/schemas/Cycles" }
        createdAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
        result: { type: object }
    Proof:
      type: object
      required: [id, theorem, status]
      properties:
        id: { type: string }
        theorem: { type: string, example: "rtl_deterministic" }
        status: { $ref: "#/components/schemas/VerificationStatus" }
        dependencies: { type: array, items: { type: string } }
        verifiedAt: { type: string, format: date-time }
        details: { type: string }
...
paths:
...
  /api/v1/devices/{id}:
    get:
      summary: Get device
      operationId: getDevice
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: ok } }
  /api/v1/devices/{id}/connect:
    post:
      summary: Connect device
      operationId: connectDevice
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: connected } }
  /api/v1/devices/{id}/disconnect:
    post:
      summary: Disconnect device
      operationId: disconnectDevice
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: disconnected } }
  /api/v1/devices/{id}/reset:
    post:
      summary: Reset device (requires hardware.reset)
      operationId: resetDevice
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: resetting }
        '403': { description: missing hardware.reset scope }
...
  /api/v1/tensor/jobs:
    post:
      summary: Create tensor job (cycles is integer)
      operationId: createTensorJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [model]
              properties:
                model: { type: string }
                inputShape: { type: array, items: { type: integer } }
      responses:
        '201':
          description: created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data: { $ref: "#/components/schemas/TensorJob" }
              example:
                id: "550e8400-e29b-41d4-a716-446655440000"
                type: "tensor.job.created"
                timestamp: "2026-01-01T00:00:00.000Z"
                source: "ortho32-api"
                correlation_id: "req-1"
                data:
                  id: "job-1"
                  status: "completed"
                  model: "ortho-llm-32"
                  cycles: 1024
                  createdAt: "2026-01-01T00:00:00.000Z"
  /api/v1/tensor/jobs/{id}:
    get:
      summary: Get tensor job
      operationId: getTensorJob
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: ok } }
  /api/v1/tensor/jobs/{id}/trace:
    get:
      summary: Get tensor trace (cycles integer)
      operationId: getTensorTrace
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: ok } }
  /api/v1/proofs/verify:
    post:
      summary: Verify proof
      operationId: verifyProof
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [theorem]
              properties:
                theorem: { type: string }
                artifact: { type: string }
      responses:
        '200':
          description: verified
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Envelope"
                  - properties:
                      data: { $ref: "#/components/schemas/Proof" }
              example:
                id: "proof-1"
                type: "proof.verified"
                timestamp: "2026-01-01T00:00:00.000Z"
                source: "ortho32-api"
                correlation_id: "req-2"
                data:
                  id: "proof-1"
                  theorem: "rtl_deterministic"
                  status: "VERIFIED"
  /api/v1/proofs/{id}:
    get:
      summary: Get proof
      operationId: getProof
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: ok } }
  /api/v1/proofs/{id}/dependencies:
    get:
      summary: Get proof dependencies
      operationId: getProofDeps
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: ok } }
  /api/v1/builds:
    post:
      summary: Create build
      operationId: createBuild
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                target: { type: string }
      responses: { '201': { description: created } }
  /api/v1/builds/{id}:
    get:
      summary: Get build
      operationId: getBuild
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: ok } }
  /api/v1/builds/{id}/logs:
    get:
      summary: Get build logs
      operationId: getBuildLogs
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: ok } }
...
  /api/v1/agents/{id}/tasks:
    post:
      summary: Create agent task
      operationId: createAgentTask
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type: { type: string }
                input: { type: object }
      responses: { '201': { description: created } }
  /api/v1/agents/{id}/tasks/{taskId}:
    get:
      summary: Get agent task
      operationId: getAgentTask
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: taskId, in: path, required: true, schema: { type: string } }
      responses: { '200': { description: ok } }
...
  /api/v1/marketplace/packages/{id}:
    get:
      summary: Get package
      operationId: getPackage
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: ok } }
  /api/v1/marketplace/packages/{id}/install:
    post:
      summary: Install package (requires marketplace.install)
      operationId: installPackage
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: installed }
        '403': { description: missing marketplace.install scope }
...